Updated playbook.yml for FF-MPEG location
Updated the playbook.yml to include the new URL for the FF-MPEG location since it's been updated. The old link no longer works and breaks the vagrant provision.

It appears to be a more generic file name with 'release' instead of a specific release specified, so hopefully this is the last time this is broken.